added BinaryParser to convert ipAddress and macAddress

+++ src/main/java/com/praca/inzynierska/gardenservicemanagement/rabbitMq/listeners/register/model/RabbitMQRegisterRequest.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class RabbitMQRegisterRequest {
     int ip;
-    int mac;
+    long mac;
     int sv;
 }
